user api to controllers and routes
