+++ client/src/utils/API.js
@@ -1,0 +1,16 @@
+import axios from "axios";
+
+export default {
+  //get all Recipes
+  getRecipes: function() {
+    return axios.get("/api/recipes");
+  },
+  //get all Ingredients
+  getIngredients: function() {
+    return axios.get("/api/ingredients");
+  },
+  //get Rating of recipe
+  getRating: function() {
+    return axios.get("/api/rating")
+  }
+};
