+++ routes/api/ratings.js
@@ -1,12 +1,13 @@
 const router = require("express").Router();
 const ratingController = require("../../controllers/ratingController");
 
+// Matches with "/api/ratings"
 router.route("/")
   .get(ratingController.getROW)
   .post(ratingController.addRating);
 
+// Matches with "/api/ratings/:id"
 router.route("/:id")
-  .get(ratingController.getRating)
   .put(ratingController.updateRating);
 
 module.exports = router;
